# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformSummary {
  activePatients: number;
  pendingAssessments: number;
  upcomingAppointments: number;
  outstandingBalance: number;
  unreadMessages: number;
}

export interface Appointment {
  id: string;
  patientName: string;
  patientId: string;
  clinicianName: string;
  type: string;
  date: string;
  time: string;
  status: string;
  location: string;
}

export interface AppointmentInput {
  type: string;
  date: string;
  time: string;
  location: string;
}

export interface Patient {
  id: string;
  name: string;
  email: string;
  initials: string;
  assessment: string;
  status: string;
  progress: number;
  lastActive: string;
}

export interface Assessment {
  id: string;
  patientName: string;
  type: string;
  stage: string;
  progress: number;
  updatedAt: string;
  clinicianName: string;
}

export interface AssessmentInput {
  patientId: string;
  type: string;
}

export interface Report {
  id: string;
  title: string;
  type: string;
  date: string;
  status: string;
  clinicianName: string;
  size: string;
}

export interface Message {
  id: string;
  sender: string;
  subject: string;
  preview: string;
  time: string;
  unread: boolean;
  direction: string;
}

export interface MessageInput {
  subject: string;
  body: string;
}

export interface Document {
  id: string;
  name: string;
  category: string;
  date: string;
  size: string;
  status: string;
}

export interface Payment {
  id: string;
  description: string;
  date: string;
  amount: number;
  status: string;
  invoice: string;
}