# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [platform]
      summary: List appointments
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [platform]
      summary: Book an assessment appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [platform]
      summary: List assessments
      responses:
        "200":
          description: Assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [platform]
      summary: Create an assessment workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Assessment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
...
  /messages:
    get:
      operationId: listMessages
      tags: [platform]
      summary: List messages
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [platform]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Appointment:
      type: object
      properties:
        id: { type: string }
        patientName: { type: string }
        patientId: { type: string }
        clinicianName: { type: string }
        type: { type: string }
        date: { type: string }
        time: { type: string }
        status: { type: string }
        location: { type: string }
      required: [id, patientName, patientId, clinicianName, type, date, time, status, location]
    AppointmentInput:
      type: object
      properties:
        type: { type: string }
        date: { type: string }
        time: { type: string }
        location: { type: string }
      required: [type, date, time, location]
...
    Assessment:
      type: object
      properties:
        id: { type: string }
        patientName: { type: string }
        type: { type: string }
        stage: { type: string }
        progress: { type: integer }
        updatedAt: { type: string }
        clinicianName: { type: string }
      required: [id, patientName, type, stage, progress, updatedAt, clinicianName]
    AssessmentInput:
      type: object
      properties:
        patientId: { type: string }
        type: { type: string }
      required: [patientId, type]
...
    Message:
      type: object
      properties:
        id: { type: string }
        sender: { type: string }
        subject: { type: string }
        preview: { type: string }
        time: { type: string }
        unread: { type: boolean }
        direction: { type: string }
      required: [id, sender, subject, preview, time, unread, direction]
    MessageInput:
      type: object
      properties:
        subject: { type: string }
        body: { type: string }
      required: [subject, body]
...