# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Appointment {
  id: string;
  patientName: string;
  patientId: string;
  clinicianName: string;
  type: string;
  date: string;
  time: string;
  status: string;
  location: string;
}