# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  type: string;
  date: string;
  time: string;
  location: string;
}